pyproject.toml
[project]
name = "zotero-assistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.70.0",
    "numpy>=2.4.4",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
]
